requirements.txt
numpy>=1.23.3
pyqtgraph>=0.13.0
PyQt6>=6.3.1
PyYAML>=6.0
pyserial>=3.5
PyVISA>=1.12.0
pulsestreamer>=1.6.0
msgpack>=1.0.4
scipy>=1.9.1
requests>=2.28.1
h5py>=3.7.0